python/requirements.txt
SQLAlchemy
pandas
psycopg2-binary
python-dotenv
